requirements.txt
requests
playwright
requests
matplotlib
dnspython
tqdm
tabulate
idna
cryptography